+++ at_location_flutter/lib/map_content/flutter_map/flutter_map.dart
@@ -7,10 +7,8 @@ import 'dart:math';
 
 import 'package:flutter/widgets.dart';
 import 'package:at_location_flutter/map_content/flutter_map/flutter_map.dart';
-import 'package:at_location_flutter/map_content/flutter_map/src/geo/crs/crs.dart';
 import 'package:at_location_flutter/map_content/flutter_map/src/map/flutter_map_state.dart';
 import 'package:at_location_flutter/map_content/flutter_map/src/map/map.dart';
-import 'package:at_location_flutter/map_content/flutter_map/src/plugins/plugin.dart';
 // ignore: import_of_legacy_library_into_null_safe
 import 'package:latlong2/latlong.dart';