+++ at_location_flutter/lib/map_content/flutter_map/src/layer/circle_layer.dart
@@ -1,7 +1,4 @@
 // ignore_for_file: prefer_void_to_null, prefer_const_constructors, avoid_unnecessary_containers, avoid_renaming_method_parameters
-
-import 'dart:ui';
-
 import 'package:flutter/widgets.dart';
 import 'package:at_location_flutter/map_content/flutter_map/flutter_map.dart';
 import 'package:at_location_flutter/map_content/flutter_map/src/map/map.dart';