+++ at_location_flutter/lib/service/location_service.dart
@@ -122,7 +122,8 @@ class LocationService {
 
       updateMyLatLng(_myData);
 
-      myLocationStream = Geolocator.getPositionStream(distanceFilter: 10)
+      myLocationStream = Geolocator.getPositionStream(
+              locationSettings: const LocationSettings(distanceFilter: 10))
           .listen((myLocation) async {
         var mylatlng = LatLng(myLocation.latitude, myLocation.longitude);