+++ at_location_flutter/lib/service/search_location_service.dart
@@ -14,6 +14,9 @@ class SearchLocationService {
   static SearchLocationService _instance = SearchLocationService._();
   factory SearchLocationService() => _instance;
 
+  final String placesUrl =
+      'https://places.ls.hereapi.com/places/v1/autosuggest';
+
   // ignore: close_sinks
   final _atLocationStreamController =
       StreamController<List<LocationModal>>.broadcast();
@@ -27,23 +30,21 @@ class SearchLocationService {
   ///
   /// Make sure that [apiKey] is passed while initialising.
   void getAddressLatLng(String address, LatLng? currentLocation) async {
-    var url;
-    // ignore: unnecessary_null_comparison
-    if (currentLocation != null) {
-      url =
-          'https://geocode.search.hereapi.com/v1/geocode?q=${address.replaceAll(RegExp(' '), '+')}&apiKey=${MixedConstants.API_KEY}&at=${currentLocation.latitude},${currentLocation.longitude}';
-    } else {
-      url =
-          'https://geocode.search.hereapi.com/v1/geocode?q=${address.replaceAll(RegExp(' '), '+')}&apiKey=${MixedConstants.API_KEY}';
-    }
+    currentLocation ??= LatLng(0, 0);
+
+    var url =
+        '$placesUrl?q=${address.replaceAll(RegExp(' '), '+')}&apiKey=${MixedConstants.API_KEY}&at=${currentLocation.latitude},${currentLocation.longitude}';
+
     var response = await http.get(Uri.parse(url));
     var addresses = jsonDecode(response.body);
-    List data = addresses['items'];
+    List data = addresses['results'];
     var share = <LocationModal>[];
     //// Removed because of nulls safety
     // for (Map ad in data ?? []) {
     for (Map ad in data) {
-      share.add(LocationModal.fromJson(ad));
+      if (ad['resultType'] == 'place') {
+        share.add(LocationModal.fromJson(ad));
+      }
     }
 
     atLocationSink.add(share);